requirements/prod.txt
Django==4.2
django-cleanup==8.1.0
django-environ==0.11.2
django-tinymce==3.7.1
python-dotenv==1.0.1
pillow==10.2.0
sorl-thumbnail==12.10.0
transliterate==1.10.2
whitenoise==6.6.0